updated AdminLayout
