+++ package.json
@@ -10,10 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@emotion/react": "^11.14.0",
-    "@emotion/styled": "^11.14.0",
     "@headlessui/react": "^2.2.0",
-    "@mui/material": "^7.0.1",
     "@tailwindcss/vite": "^4.0.17",
     "@tanstack/react-query": "^5.69.0",
     "antd": "^5.24.6",
